+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 Added
 ~~~~~
 - ``Entity.top`` property to always return the first child in the list
+- ``ContactMap.find`` function accepts ``strict`` keyword argument to find contact pairs with both residues in ``register``
 
 [0.8.3]
 -------